requirements.txt
d2to1==0.2.11
pbr==0.5.21
SQLAlchemy==0.7.9
sqlalchemy-migrate==0.7.2
anyjson==0.3.3
iso8601==0.1.8
netaddr==0.7.5
argparse==1.2.1
eventlet==0.13.0
python-keystoneclient>=0.3.2
python-designateclient==0.0.4
pecan==0.3.2
WSME==0.5b5
MySQL-python==1.2.3c1
oslo.config==1.2.0
httplib2==0.7.7
six==1.4.1
scapy-real==2.2.0-dev
tailer==0.3
python-memcached==1.43
